pyproject.toml
[project]
name = "sanset-backend"
version = "0.1.0"
description = "Sanset FastAPI Backend - AI-powered e-commerce platform"
authors = ["Sanset Team"]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "email-validator>=2.1.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "apscheduler>=3.10.0",
    "python-dateutil>=2.8.0",
    "psycopg2-binary>=2.9.9",
]
